+++ postgres/src/lib.rs
@@ -51,15 +51,15 @@ use tokio_postgres::{
 };
 
 /// A type alias for using `deadpool::Pool` with `tokio_postgres`
-pub type Pool = deadpool::Pool<ClientWrapper, tokio_postgres::Error>;
+pub type Pool = deadpool::managed::Pool<ClientWrapper, tokio_postgres::Error>;
 
 /// A type alias for using `deadpool::PoolError` with `tokio_postgres`
-pub type PoolError = deadpool::PoolError<tokio_postgres::Error>;
+pub type PoolError = deadpool::managed::PoolError<tokio_postgres::Error>;
 
 /// A type alias for using `deadpool::Object` with `tokio_postgres`
-pub type Client = deadpool::Object<ClientWrapper, tokio_postgres::Error>;
+pub type Client = deadpool::managed::Object<ClientWrapper, tokio_postgres::Error>;
 
-type RecycleResult = deadpool::RecycleResult<Error>;
+type RecycleResult = deadpool::managed::RecycleResult<Error>;
 
 /// The manager for creating and recyling postgresql connections
 pub struct Manager<T: MakeTlsConnect<Socket>> {
@@ -78,7 +78,7 @@ impl<T: MakeTlsConnect<Socket>> Manager<T> {
 }
 
 #[async_trait]
-impl<T> deadpool::Manager<ClientWrapper, Error> for Manager<T>
+impl<T> deadpool::managed::Manager<ClientWrapper, Error> for Manager<T>
 where
     T: MakeTlsConnect<Socket> + Clone + Sync + Send + 'static,
     T::Stream: Sync + Send,